+++ Airship/AirshipDebug/Source/DebugComponent.swift
@@ -1,6 +1,7 @@
 /* Copyright Airship and Contributors */
 
 import Foundation
+@preconcurrency import UserNotifications
 
 #if canImport(AirshipCore)
 import AirshipCore
@@ -13,5 +14,22 @@ final class DebugComponent : AirshipComponent, AirshipPushableComponent {
     init(debugManager: AirshipDebugManager) {
         self.debugManager = debugManager
     }
+
+    @MainActor
+    func receivedRemoteNotification(
+        _ notification: AirshipCore.AirshipJSON
+    ) async -> AirshipCore.UABackgroundFetchResult {
+        return await self.debugManager.receivedRemoteNotification(notification)
+    }
+
+#if !os(tvOS)
+    @MainActor
+    func receivedNotificationResponse(
+        _ response: UNNotificationResponse
+    ) async {
+        return await self.debugManager.receivedNotificationResponse(response)
+    }
+#endif
+
 }